backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
supabase==2.3.0
psycopg2-binary==2.9.9

# OpenAI and AI
openai==1.10.0
httpx==0.26.0

# Temporal workflow
temporalio==1.5.1

# Geospatial
geoalchemy2==0.14.3
numpy==1.26.3

# Memory and caching
redis==5.0.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Memory and AI Agents
mem0ai==0.1.5

# End-to-End Encryption
PyNaCl==1.5.0
argon2-cffi==23.1.0
